pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "execucode"
version = "0.1.0"
description = "Conversational execution-aware code optimization environment for OpenEnv."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115",
    "openai>=1.0",
    "openenv-core>=0.2.3",
    "pydantic>=2.0",
    "uvicorn[standard]>=0.30",
]

[tool.setuptools]
packages = ["execucode", "execucode.server"]

[tool.setuptools.package-dir]
execucode = "."
"execucode.server" = "server"
